refactor!: index stats to handle error and bunch of bugs (#1828)
BREAKING CHANGE: removed single-purpose stats API from public API and
refactored `DatasetIndexExt` to `lance-index`.

Also, fixed a few places that `unwrap()` results.
