+++ rust/lance-index/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 Lance Developers.
+// Copyright 2024 Lance Developers.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,7 +24,9 @@ use roaring::RoaringBitmap;
 
 pub mod optimize;
 pub mod scalar;
+pub mod traits;
 pub mod vector;
+pub use crate::traits::*;
 
 pub const INDEX_FILE_NAME: &str = "index.idx";
 
@@ -34,16 +36,21 @@ pub mod pb {
 }
 
 /// Generic methods common across all types of secondary indices
+///
 #[async_trait]
 pub trait Index: Send + Sync {
     /// Cast to [Any].
     fn as_any(&self) -> &dyn Any;
+
     /// Cast to [Index]
     fn as_index(self: Arc<Self>) -> Arc<dyn Index>;
+
     /// Retrieve index statistics as a JSON string
     fn statistics(&self) -> Result<String>;
+
     /// Get the type of the index
     fn index_type(&self) -> IndexType;
+
     /// Read through the index and determine which fragment ids are covered by the index
     ///
     /// This is a kind of slow operation.  It's better to use the fragment_bitmap.  This
@@ -68,4 +75,8 @@ impl std::fmt::Display for IndexType {
             Self::Vector => write!(f, "Vector"),
         }
     }
+}
+
+pub trait IndexParams: Send + Sync {
+    fn as_any(&self) -> &dyn Any;
 }
