+++ rust/lance/benches/ivf_pq.rs
@@ -17,14 +17,14 @@ use std::sync::Arc;
 use arrow_array::{FixedSizeListArray, RecordBatch, RecordBatchIterator};
 use arrow_schema::{DataType, Field, FieldRef, Schema};
 use criterion::{criterion_group, criterion_main, Criterion};
+
 use lance::{
     arrow::*,
     dataset::{WriteMode, WriteParams},
-    index::{vector::VectorIndexParams, DatasetIndexExt},
+    index::vector::VectorIndexParams,
     Dataset,
 };
-
-use lance_index::IndexType;
+use lance_index::{DatasetIndexExt, IndexType};
 use lance_linalg::distance::MetricType;
 use lance_testing::datagen::generate_random_array;
 #[cfg(target_os = "linux")]
