+++ rust/lance/src/index.rs
@@ -15,19 +15,22 @@
 //! Secondary Index
 //!
 
-use std::any::Any;
 use std::collections::HashMap;
 use std::sync::Arc;
 
 use arrow_schema::DataType;
 use async_trait::async_trait;
-use lance_core::io::{read_message, read_message_from_buf, read_metadata_offset, Reader};
+use lance_core::format::Fragment;
+use lance_core::io::{
+    read_message, read_message_from_buf, read_metadata_offset, reader::read_manifest_indexes,
+    Reader,
+};
+use lance_index::pb::index::Implementation;
 use lance_index::scalar::expression::IndexInformationProvider;
 use lance_index::scalar::lance_format::LanceIndexStore;
 use lance_index::scalar::ScalarIndex;
-use lance_index::{optimize::OptimizeOptions, pb::index::Implementation};
-use lance_index::{pb, Index, IndexType, INDEX_FILE_NAME};
-use log::warn;
+pub use lance_index::IndexParams;
+use lance_index::{pb, DatasetIndexExt, Index, IndexType, INDEX_FILE_NAME};
 use snafu::{location, Location};
 use tracing::instrument;
 use uuid::Uuid;
@@ -38,6 +41,7 @@ pub(crate) mod prefilter;
 pub mod scalar;
 pub mod vector;
 
+use crate::dataset::index::unindexed_fragments;
 use crate::dataset::transaction::{Operation, Transaction};
 use crate::format::Index as IndexMetadata;
 use crate::index::append::append_index;
@@ -54,10 +58,6 @@ pub trait IndexBuilder {
     fn index_type() -> IndexType;
 
     async fn build(&self) -> Result<()>;
-}
-
-pub trait IndexParams: Send + Sync {
-    fn as_any(&self) -> &dyn Any;
 }
 
 pub(crate) async fn remap_index(
@@ -129,39 +129,6 @@ impl IndexInformationProvider for ScalarIndexInfo {
     fn get_index(&self, col: &str) -> Option<&DataType> {
         self.indexed_columns.get(col)
     }
-}
-
-/// Extends Dataset with secondary index.
-#[async_trait]
-pub trait DatasetIndexExt {
-    /// Create indices on columns.
-    ///
-    /// Upon finish, a new dataset version is generated.
-    ///
-    /// Parameters:
-    ///
-    ///  - `columns`: the columns to build the indices on.
-    ///  - `index_type`: specify [`IndexType`].
-    ///  - `name`: optional index name. Must be unique in the dataset.
-    ///            if not provided, it will auto-generate one.
-    ///  - `params`: index parameters.
-    ///  - `replace`: replace the existing index if it exists.
-    async fn create_index(
-        &mut self,
-        columns: &[&str],
-        index_type: IndexType,
-        name: Option<String>,
-        params: &dyn IndexParams,
-        replace: bool,
-    ) -> Result<()>;
-
-    /// Optimize indices.
-    ///
-    /// Parameters
-    /// ----------
-    /// options: OptimizeOptions
-    ///     Options for optimizing indices.
-    async fn optimize_indices(&mut self, options: OptimizeOptions) -> Result<()>;
 }
 
 async fn open_index_proto(dataset: &Dataset, reader: &dyn Reader) -> Result<pb::Index> {
@@ -285,6 +252,27 @@ impl DatasetIndexExt for Dataset {
         Ok(())
     }
 
+    async fn load_indices(&self) -> Result<Vec<IndexMetadata>> {
+        let manifest_file = self.manifest_file(self.version().version).await?;
+        read_manifest_indexes(&self.object_store, &manifest_file, &self.manifest).await
+    }
+
+    async fn load_scalar_index_for_column(&self, col: &str) -> Result<Option<IndexMetadata>> {
+        Ok(self
+            .load_indices()
+            .await?
+            .into_iter()
+            .filter(|idx| idx.fields.len() == 1)
+            .find(|idx| {
+                let field = self.schema().field_by_id(idx.fields[0]);
+                if let Some(field) = field {
+                    field.name == col
+                } else {
+                    false
+                }
+            }))
+    }
+
     #[instrument(skip_all)]
     async fn optimize_indices(&mut self, options: OptimizeOptions) -> Result<()> {
         let dataset = Arc::new(self.clone());
@@ -367,6 +355,53 @@ impl DatasetIndexExt for Dataset {
 
         self.manifest = Arc::new(new_manifest);
         Ok(())
+    }
+
+    async fn index_statistics(&self, index_name: &str) -> Result<Option<String>> {
+        let index_uuid = self
+            .load_index_by_name(index_name)
+            .await?
+            .map(|idx| idx.uuid.to_string());
+
+        if let Some(index_uuid) = index_uuid {
+            let index_statistics = self
+                .open_generic_index("vector", &index_uuid)
+                .await?
+                .statistics()?;
+            Ok(Some(index_statistics))
+        } else {
+            Ok(None)
+        }
+    }
+
+    async fn count_unindexed_rows(&self, index_uuid: &str) -> Result<Option<usize>> {
+        let index = self.load_index(index_uuid).await?;
+
+        if let Some(index) = index {
+            let unindexed_frags = unindexed_fragments(&index, self).await?;
+            let unindexed_rows = unindexed_frags
+                .iter()
+                .map(Fragment::num_rows)
+                // sum the number of rows in each fragment if no fragment returned None from row_count
+                .try_fold(0, |a, b| b.map(|b| a + b).ok_or(()));
+
+            Ok(unindexed_rows.ok())
+        } else {
+            Ok(None)
+        }
+    }
+
+    async fn count_indexed_rows(&self, index_uuid: &str) -> Result<Option<usize>> {
+        let count_rows = self.count_rows();
+        let count_unindexed_rows = self.count_unindexed_rows(index_uuid);
+
+        let (count_rows, count_unindexed_rows) =
+            futures::try_join!(count_rows, count_unindexed_rows)?;
+
+        match count_unindexed_rows {
+            Some(count_unindexed_rows) => Ok(Some(count_rows - count_unindexed_rows)),
+            None => Ok(None),
+        }
     }
 }
 
@@ -539,5 +574,90 @@ mod tests {
             )
             .await
             .is_err());
+    }
+
+    #[tokio::test]
+    async fn test_count_index_rows() {
+        let test_dir = tempdir().unwrap();
+        let dimensions = 16;
+        let column_name = "vec";
+        let field = Field::new(
+            column_name,
+            DataType::FixedSizeList(
+                Arc::new(Field::new("item", DataType::Float32, true)),
+                dimensions,
+            ),
+            false,
+        );
+        let schema = Arc::new(Schema::new(vec![field]));
+
+        let float_arr = generate_random_array(512 * dimensions as usize);
+
+        let vectors =
+            arrow_array::FixedSizeListArray::try_new_from_values(float_arr, dimensions).unwrap();
+
+        let record_batch = RecordBatch::try_new(schema.clone(), vec![Arc::new(vectors)]).unwrap();
+
+        let reader =
+            RecordBatchIterator::new(vec![record_batch].into_iter().map(Ok), schema.clone());
+
+        let test_uri = test_dir.path().to_str().unwrap();
+        let mut dataset = Dataset::write(reader, test_uri, None).await.unwrap();
+        dataset.validate().await.unwrap();
+
+        // Make sure it returns None if there's no index with the passed identifier
+        assert_eq!(dataset.count_unindexed_rows("bad_id").await.unwrap(), None);
+        assert_eq!(dataset.count_indexed_rows("bad_id").await.unwrap(), None);
+
+        // Create an index
+        let params = VectorIndexParams::ivf_pq(10, 8, 2, false, MetricType::L2, 10);
+        dataset
+            .create_index(
+                &[column_name],
+                IndexType::Vector,
+                Some("vec_idx".into()),
+                &params,
+                true,
+            )
+            .await
+            .unwrap();
+
+        let index = dataset
+            .load_index_by_name("vec_idx")
+            .await
+            .unwrap()
+            .unwrap();
+        let index_uuid = &index.uuid.to_string();
+
+        // Make sure there are no unindexed rows
+        assert_eq!(
+            dataset.count_unindexed_rows(index_uuid).await.unwrap(),
+            Some(0)
+        );
+        assert_eq!(
+            dataset.count_indexed_rows(index_uuid).await.unwrap(),
+            Some(512)
+        );
+
+        // Now we'll append some rows which shouldn't be indexed and see the
+        // count change
+        let float_arr = generate_random_array(512 * dimensions as usize);
+        let vectors =
+            arrow_array::FixedSizeListArray::try_new_from_values(float_arr, dimensions).unwrap();
+
+        let record_batch = RecordBatch::try_new(schema.clone(), vec![Arc::new(vectors)]).unwrap();
+
+        let reader = RecordBatchIterator::new(vec![record_batch].into_iter().map(Ok), schema);
+        dataset.append(reader, None).await.unwrap();
+
+        // Make sure the new rows are not indexed
+        assert_eq!(
+            dataset.count_unindexed_rows(index_uuid).await.unwrap(),
+            Some(512)
+        );
+        assert_eq!(
+            dataset.count_indexed_rows(index_uuid).await.unwrap(),
+            Some(512)
+        );
     }
 }
